Split out transaction functions from storage
