+++ packages/std/src/storage.rs
@@ -1,15 +1,10 @@
-#[cfg(feature = "iterator")]
-use std::cmp::Ordering;
 use std::collections::BTreeMap;
-#[cfg(feature = "iterator")]
-use std::iter::Peekable;
 #[cfg(feature = "iterator")]
 use std::ops::Bound;
 
-use crate::errors::Result;
-use crate::traits::{Api, Extern, ReadonlyStorage, Storage};
 #[cfg(feature = "iterator")]
 use crate::traits::{KVPair, Sort};
+use crate::traits::{ReadonlyStorage, Storage};
 
 #[derive(Default)]
 pub struct MemoryStorage {
@@ -57,188 +52,6 @@ impl Storage for MemoryStorage {
     }
 }
 
-pub struct StorageTransaction<'a, S: ReadonlyStorage> {
-    /// read-only access to backing storage
-    storage: &'a S,
-    /// these are local changes not flushed to backing storage
-    local_state: MemoryStorage,
-    /// a log of local changes not yet flushed to backing storage
-    rep_log: RepLog,
-}
-
-pub struct RepLog {
-    /// this is a list of changes to be written to backing storage upon commit
-    ops_log: Vec<Op>,
-}
-
-impl RepLog {
-    fn new() -> Self {
-        RepLog { ops_log: vec![] }
-    }
-
-    /// appends an op to the list of changes to be applied upon commit
-    fn append(&mut self, op: Op) {
-        self.ops_log.push(op);
-    }
-
-    /// applies the stored list of `Op`s to the provided `Storage`
-    pub fn commit<S: Storage>(self, storage: &mut S) {
-        for op in self.ops_log {
-            op.apply(storage);
-        }
-    }
-}
-
-enum Op {
-    /// represents the `Set` operation for setting a key-value pair in storage
-    Set { key: Vec<u8>, value: Vec<u8> },
-}
-
-impl Op {
-    /// applies this `Op` to the provided storage
-    pub fn apply<S: Storage>(&self, storage: &mut S) {
-        match self {
-            Op::Set { key, value } => storage.set(&key, &value),
-        }
-    }
-}
-
-impl<'a, S: ReadonlyStorage> StorageTransaction<'a, S> {
-    pub fn new(storage: &'a S) -> Self {
-        StorageTransaction {
-            storage,
-            local_state: MemoryStorage::new(),
-            rep_log: RepLog::new(),
-        }
-    }
-
-    /// prepares this transaction to be committed to storage
-    pub fn prepare(self) -> RepLog {
-        self.rep_log
-    }
-
-    /// rollback will consume the checkpoint and drop all changes (no really needed, going out of scope does the same, but nice for clarity)
-    pub fn rollback(self) {}
-}
-
-impl<'a, S: ReadonlyStorage> ReadonlyStorage for StorageTransaction<'a, S> {
-    fn get(&self, key: &[u8]) -> Option<Vec<u8>> {
-        match self.local_state.get(key) {
-            Some(val) => Some(val),
-            None => self.storage.get(key),
-        }
-    }
-
-    #[cfg(feature = "iterator")]
-    /// range allows iteration over a set of keys, either forwards or backwards
-    /// uses standard rust range notation, and eg db.range(b"foo"..b"bar") also works reverse
-    fn range(
-        &self,
-        start: Option<&[u8]>,
-        end: Option<&[u8]>,
-        order: Sort,
-    ) -> Box<dyn Iterator<Item = KVPair>> {
-        let local = self.local_state.range(start, end, order);
-        let base = self.storage.range(start, end, order);
-        let merged = MergeOrdered::new(local, base, order);
-        Box::new(merged)
-    }
-}
-
-impl<'a, S: ReadonlyStorage> Storage for StorageTransaction<'a, S> {
-    fn set(&mut self, key: &[u8], value: &[u8]) {
-        let op = Op::Set {
-            key: key.to_vec(),
-            value: value.to_vec(),
-        };
-        op.apply(&mut self.local_state);
-        self.rep_log.append(op);
-    }
-}
-
-#[cfg(feature = "iterator")]
-struct MergeOrdered<L, R>
-where
-    L: Iterator<Item = R::Item>,
-    R: Iterator,
-{
-    left: Peekable<L>,
-    right: Peekable<R>,
-    order: Sort,
-}
-
-#[cfg(feature = "iterator")]
-impl<L, R> MergeOrdered<L, R>
-where
-    L: Iterator<Item = R::Item>,
-    R: Iterator,
-{
-    fn new(left: L, right: R, order: Sort) -> Self {
-        MergeOrdered {
-            left: left.peekable(),
-            right: right.peekable(),
-            order,
-        }
-    }
-}
-
-#[cfg(feature = "iterator")]
-impl<L, R> Iterator for MergeOrdered<L, R>
-where
-    L: Iterator<Item = R::Item>,
-    R: Iterator,
-    L::Item: Ord,
-{
-    type Item = L::Item;
-
-    fn next(&mut self) -> Option<L::Item> {
-        match (self.left.peek(), self.right.peek()) {
-            (Some(l), Some(r)) => {
-                let order = match self.order {
-                    Sort::Ascending => l.cmp(r),
-                    Sort::Descending => r.cmp(l),
-                };
-                match order {
-                    Ordering::Less => self.left.next(),
-                    Ordering::Equal => self.left.next(),
-                    Ordering::Greater => self.right.next(),
-                }
-            }
-            (Some(_), None) => self.left.next(),
-            (None, Some(_)) => self.right.next(),
-            (None, None) => None,
-        }
-    }
-}
-
-pub fn transactional<S: Storage, T>(
-    storage: &mut S,
-    tx: &dyn Fn(&mut StorageTransaction<S>) -> Result<T>,
-) -> Result<T> {
-    let mut stx = StorageTransaction::new(storage);
-    let res = tx(&mut stx)?;
-    stx.prepare().commit(storage);
-    Ok(res)
-}
-
-pub fn transactional_deps<S: Storage, A: Api, T>(
-    deps: &mut Extern<S, A>,
-    tx: &dyn Fn(&mut Extern<StorageTransaction<S>, A>) -> Result<T>,
-) -> Result<T> {
-    let c = StorageTransaction::new(&deps.storage);
-    let mut stx_deps = Extern {
-        storage: c,
-        api: deps.api,
-    };
-    let res = tx(&mut stx_deps);
-    if res.is_ok() {
-        stx_deps.storage.prepare().commit(&mut deps.storage);
-    } else {
-        stx_deps.storage.rollback();
-    }
-    res
-}
-
 #[cfg(test)]
 #[cfg(feature = "iterator")]
 // iterator_test_suite takes a storage, adds data and runs iterator tests
@@ -246,7 +59,7 @@ pub fn transactional_deps<S: Storage, A: Api, T>(
 // (this allows us to test StorageTransaction and other wrapped storage better)
 //
 // designed to be imported by other modules
-fn iterator_test_suite<S: Storage>(store: &mut S) {
+pub(crate) fn iterator_test_suite<S: Storage>(store: &mut S) {
     // ensure we had previously set "foo" = "bar"
     assert_eq!(store.get(b"foo"), Some(b"bar".to_vec()));
     assert_eq!(store.range(None, None, Sort::Ascending).count(), 1);
@@ -327,7 +140,6 @@ fn iterator_test_suite<S: Storage>(store: &mut S) {
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::errors::Unauthorized;
 
     #[test]
     fn memory_storage_get_and_set() {
@@ -344,109 +156,5 @@ mod test {
         let mut store = MemoryStorage::new();
         store.set(b"foo", b"bar");
         iterator_test_suite(&mut store);
-    }
-
-    #[test]
-    #[cfg(feature = "iterator")]
-    fn storage_transaction_iterator_empty_base() {
-        let mut base = MemoryStorage::new();
-        let mut check = StorageTransaction::new(&mut base);
-        check.set(b"foo", b"bar");
-        iterator_test_suite(&mut check);
-    }
-
-    #[test]
-    #[cfg(feature = "iterator")]
-    fn storage_transaction_iterator_with_base_data() {
-        let mut base = MemoryStorage::new();
-        base.set(b"foo", b"bar");
-        let mut check = StorageTransaction::new(&mut base);
-        iterator_test_suite(&mut check);
-    }
-
-    #[test]
-    fn commit_writes_through() {
-        let mut base = MemoryStorage::new();
-        base.set(b"foo", b"bar");
-
-        let mut check = StorageTransaction::new(&base);
-        assert_eq!(check.get(b"foo"), Some(b"bar".to_vec()));
-        check.set(b"subtx", b"works");
-        check.prepare().commit(&mut base);
-
-        assert_eq!(base.get(b"subtx"), Some(b"works".to_vec()));
-    }
-
-    #[test]
-    fn storage_remains_readable() {
-        let mut base = MemoryStorage::new();
-        base.set(b"foo", b"bar");
-
-        let mut stxn1 = StorageTransaction::new(&base);
-
-        assert_eq!(stxn1.get(b"foo"), Some(b"bar".to_vec()));
-
-        stxn1.set(b"subtx", b"works");
-        assert_eq!(stxn1.get(b"subtx"), Some(b"works".to_vec()));
-
-        // Can still read from base, txn is not yet committed
-        assert_eq!(base.get(b"subtx"), None);
-
-        stxn1.prepare().commit(&mut base);
-        assert_eq!(base.get(b"subtx"), Some(b"works".to_vec()));
-    }
-
-    #[test]
-    fn rollback_has_no_effect() {
-        let mut base = MemoryStorage::new();
-        base.set(b"foo", b"bar");
-
-        let mut check = StorageTransaction::new(&mut base);
-        assert_eq!(check.get(b"foo"), Some(b"bar".to_vec()));
-        check.set(b"subtx", b"works");
-        check.rollback();
-
-        assert_eq!(base.get(b"subtx"), None);
-    }
-
-    #[test]
-    fn ignore_same_as_rollback() {
-        let mut base = MemoryStorage::new();
-        base.set(b"foo", b"bar");
-
-        let mut check = StorageTransaction::new(&mut base);
-        assert_eq!(check.get(b"foo"), Some(b"bar".to_vec()));
-        check.set(b"subtx", b"works");
-
-        assert_eq!(base.get(b"subtx"), None);
-    }
-
-    #[test]
-    fn transactional_works() {
-        let mut base = MemoryStorage::new();
-        base.set(b"foo", b"bar");
-
-        // writes on success
-        let res: Result<i32> = transactional(&mut base, &|store| {
-            // ensure we can read from the backing store
-            assert_eq!(store.get(b"foo"), Some(b"bar".to_vec()));
-            // we write in the Ok case
-            store.set(b"good", b"one");
-            Ok(5)
-        });
-        assert_eq!(5, res.unwrap());
-        assert_eq!(base.get(b"good"), Some(b"one".to_vec()));
-
-        // rejects on error
-        let res: Result<i32> = transactional(&mut base, &|store| {
-            // ensure we can read from the backing store
-            assert_eq!(store.get(b"foo"), Some(b"bar".to_vec()));
-            assert_eq!(store.get(b"good"), Some(b"one".to_vec()));
-            // we write in the Error case
-            store.set(b"bad", b"value");
-            Unauthorized.fail()
-        });
-        assert!(res.is_err());
-        assert_eq!(base.get(b"bad"), None);
     }
 }
